# doc/src/main/java/com/example/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OnlineShop API",
    "description": "java onlineshop",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://127.0.0.1",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/panel/user/update_profile": {
      "put": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "editProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/user/change_pass": {
      "put": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "changePassword",
        "parameters": [
          {
            "name": "dto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChangePassDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/user/change_pass/admin": {
      "put": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "changePasswordByAdmin",
        "parameters": [
          {
            "name": "userDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UserDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/slider/swap_up/{id}": {
      "put": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "swapUp",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/slider/swap_down/{id}": {
      "put": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "swapDown",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/nav/swap_down/{id}": {
      "put": {
        "tags": [
          "nav-panel-controller"
        ],
        "operationId": "swapDown_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/file/upload": {
      "post": {
        "tags": [
          "file-panel-controller"
        ],
        "operationId": "upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseFileDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/user": {
      "get": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListUserDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/user/{id}": {
      "get": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "getById_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseUserDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-panel-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/panel/slider": {
      "get": {
        "tags": [
          "slider-panel-controller"
        ],
        "operationId": "getAll_2",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListSliderDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/invoice/user/{uid}": {
      "get": {
        "tags": [
          "invoice-panel-controller"
        ],
        "operationId": "getAllById",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponseListInvoiceDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/panel/file": {
      "get": {
        "tags": [
          "file-panel-controller"
        ],
        "operationId": "getAll_7",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIPanelResponseListFileDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "APIResponseUserDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "CustomerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "tel": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          }
        }
      },
      "PermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            },
            "uniqueItems": true
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "registerDate": {
            "type": "string",
            "format": "date-time"
          },
          "enable": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleDto"
            },
            "uniqueItems": true
          },
          "customerDto": {
            "$ref": "#/components/schemas/CustomerDto"
          }
        }
      },
      "ChangePassDto": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "newPassword2": {
            "type": "string"
          }
        }
      },
      "APIResponseBoolean": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "boolean"
          }
        }
      },
      "File": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SliderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "orderNumber": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "$ref": "#/components/schemas/File"
          }
        }
      },
...
      "SizeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "ColorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "hex": {
            "type": "string"
          }
        }
      },
      "FileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          },
          "extension": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "enable": {
            "type": "boolean"
          },
          "exist": {
            "type": "boolean"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          },
          "visitCount": {
            "type": "integer",
            "format": "int64"
          },
          "addDate": {
            "type": "string",
            "format": "date-time"
          },
          "image": {
            "$ref": "#/components/schemas/FileDto"
          },
          "colors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            },
            "uniqueItems": true
          },
          "sizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SizeDto"
            },
            "uniqueItems": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategoryDto"
          }
        }
      },
...
      "APIResponseFileDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "$ref": "#/components/schemas/FileDto"
          }
        }
      },
...
      "APIPanelResponseListUserDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "APIPanelResponseListSliderDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SliderDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "payedDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Payed",
              "InProgress"
            ]
          },
          "totalAmount": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            }
          }
        }
      },
      "InvoiceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "payedDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "Payed",
              "InProgress"
            ]
          },
          "totalAmount": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemDto"
            }
          }
        }
      },
      "InvoiceItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "invoice": {
            "$ref": "#/components/schemas/Invoice"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "size": {
            "$ref": "#/components/schemas/Size"
          },
          "color": {
            "$ref": "#/components/schemas/Color"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "InvoiceItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "invoice": {
            "$ref": "#/components/schemas/Invoice"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDto"
          },
          "size": {
            "$ref": "#/components/schemas/SizeDto"
          },
          "color": {
            "$ref": "#/components/schemas/ColorDto"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "Size": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "registerDate": {
            "type": "string",
            "format": "date-time"
          },
          "enable": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "uniqueItems": true
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          }
        }
      },
      "APIResponseListInvoiceDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceDto"
            }
          }
        }
      },
      "APIPanelResponseListFileDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Success",
              "Error",
              "Failed",
              "Redirect",
              "Forbidden",
              "AccessDenied"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}